[frontend] fix Analyses chip in Knowledge search (#13303)
