+++ opencti-platform/opencti-front/src/private/components/SearchBulkContainer.tsx
@@ -1,28 +1,21 @@
-import React, { ChangeEvent, MouseEvent, useState } from 'react';
+import React, { ChangeEvent, useState } from 'react';
 import Grid from '@mui/material/Grid';
 import TextField from '@mui/material/TextField';
 import Box from '@mui/material/Box';
 import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
 import { isEmpty } from 'ramda';
-import Chip from '@mui/material/Chip';
-import { useNavigate } from 'react-router-dom';
-import { useTheme } from '@mui/material/styles';
 import SearchBulkUnknownEntities from './SearchBulkUnknownEntities';
 import { useFormatter } from '../../components/i18n';
 import Breadcrumbs from '../../components/Breadcrumbs';
 import useConnectedDocumentModifier from '../../utils/hooks/useConnectedDocumentModifier';
 import SearchBulk, { BULK_SEARCH_LOCAL_STORAGE_KEY } from './SearchBulk';
 import DataTableWithoutFragment from '../../components/dataGrid/DataTableWithoutFragment';
-import { resolveLink } from '../../utils/Entity';
-import { typesWithNoAnalysesTab } from '../../utils/hooks/useAttributes';
 import { DataTableProps } from '../../components/dataGrid/dataTableTypes';
 import useDebounceCallback from '../../utils/hooks/useDebounceCallback';
 
 const SearchBulkContainer = () => {
-  const { t_i18n, n } = useFormatter();
-  const theme = useTheme();
-  const navigate = useNavigate();
+  const { t_i18n } = useFormatter();
   const { setTitle } = useConnectedDocumentModifier();
 
   setTitle(t_i18n('Bulk Search'));
@@ -74,50 +67,6 @@ const SearchBulkContainer = () => {
     },
     analyses: {
       percentWidth: 7,
-      id: 'analyses',
-      label: 'Analyses',
-      isSortable: false,
-      render: ({ id, entity_type, containersNumber }) => {
-        const analysesNumber = containersNumber?.total;
-        const link = `${resolveLink(entity_type)}/${id}`;
-        const linkAnalyses = `${link}/analyses`;
-        const analysesChipStyle: React.CSSProperties = {
-          fontSize: 13,
-          lineHeight: '12px',
-          height: 20,
-          textTransform: 'uppercase',
-          borderRadius: 4,
-        };
-        return (
-          <>
-            {typesWithNoAnalysesTab.includes(entity_type)
-              ? (
-                  <Chip
-                    style={analysesChipStyle}
-                    label={n(analysesNumber)}
-                  />
-                )
-              : (
-                  <Chip
-                    sx={{
-                      ...analysesChipStyle,
-                      cursor: 'pointer',
-                      '&:hover': {
-                        backgroundColor: theme.palette.primary.main,
-                      },
-                    }}
-                    label={n(analysesNumber)}
-                    onClick={(e: MouseEvent<HTMLDivElement>) => {
-                      e.preventDefault();
-                      e.stopPropagation();
-                      navigate(linkAnalyses);
-                    }}
-                  />
-                )
-            }
-          </>
-        );
-      },
     },
     objectMarking: {},
   };
